Add new license implementation on Login and Register screen
- LEARNER-4959
Parts of the license text have been made clickable by applying UrlSpans
on them. The URI attached to each link within the text is a custom made
URI that is broadcasted upon a link's click and received by one of our
activity (WebViewActivity) using an intent filter defined in
AndroidManifest.xml for the said activity.